Added support for limiting the number of snapshots

+++ extra/vm_backups.py
@@ -48,15 +48,13 @@ class VmSnapshot():
             
             # Generate list of snapshots to delete
             vm_zfs_snapshots_to_delete = vm_zfs_snapshot_list.copy()
-            for zfs_snapshot in range(0, snapshots_to_keep):
-                try:
+            if len(vm_zfs_snapshots_to_delete) > 0:
+                for zfs_snapshot in range(0, snapshots_to_keep):
                     del vm_zfs_snapshots_to_delete[-1]
-                except:
-                    print("VM " + self.vmname + " doesn't have any snapshots to delete")
-            # Remove the old snapshots
-            for vm_zfs_snapshot_to_delete in vm_zfs_snapshots_to_delete:
-                command = "zfs destroy " + vm_zfs_snapshot_to_delete
-                print(command)
+                # Remove the old snapshots
+                for vm_zfs_snapshot_to_delete in vm_zfs_snapshots_to_delete:
+                    command = "zfs destroy " + vm_zfs_snapshot_to_delete
+                    print(command)
 
         else:
             command = "zfs snapshot " + zfs_dataset + "@" + snapshot_name
